requirements.txt
python-telegram-bot>=20.0
cloudflare>=2.19.0
python-dotenv>=1.0.0